Re-structure code organization
