+++ entities/categoria.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+class Category:
+    DESPORTO = 'desporto'
+    CINEMA = 'cinema'
+    MUSICA = 'musica'
+    TEATRO = 'teatro'
+    FESTA = 'festa'
+    WORKSHOP = 'workshop'
+    OUTRO = 'outro'
