+++ entities/evento.py
@@ -1,0 +1,7 @@
+class Evento:
+    name = ""
+    categories = []
+    thumbnail = ""
+    location = ""
+    title = ""
+    date = ""
